python_ecs/requirements.txt
boto3==1.34.162
botocore==1.34.162
Pillow==10.4.0
